requirements.txt
# Azure Functions for Python
azure-functions>=1.20.0

# Azure SDK dependencies
azure-identity>=1.16.0
azure-keyvault-secrets>=4.8.0
azure-storage-file-datalake>=12.14.0

# For testing
pytest>=7.4.0
pytest-mock>=3.11.1
pytest-cov>=4.1.0
